test(playback): declare probe test platform requirements
